+++ src/Factory/ContentStructureFactory.php
@@ -12,6 +12,9 @@ use eZ\Publish\API\Repository\Exceptions\NotFoundException;
 
 final class ContentStructureFactory
 {
+    public const MODE_INSERT_OR_UPDATE = 1;
+    public const MODE_INSERT_ONLY = 2;
+    public const MODE_UPDATE_ONLY = 3;
     /**
      * @var ContentService
      */
@@ -28,27 +31,32 @@ final class ContentStructureFactory
     }
 
     /**
-     * @param array  $data
+     * @param array $data
      * @param string $remoteId
      * @param string $language
      * @param string $contentType
-     * @param mixed  $parentLocations
-     *
-     * @return ContentStructure
+     * @param mixed $parentLocations
+     * @param int $mode One of the constant ContentStructureFactory::MODE_*
+     * @return false|ContentStructure
      *
      * @throws \CodeRhapsodie\EzDataflowBundle\Exception\InvalidArgumentTypeException
      * @throws \eZ\Publish\API\Repository\Exceptions\UnauthorizedException
      */
-    public function transform(array $data, string $remoteId, string $language, string $contentType, $parentLocations): ContentStructure
+    public function transform(array $data, string $remoteId, string $language, string $contentType, $parentLocations, int $mode = ContentStructureFactory::MODE_INSERT_OR_UPDATE)
     {
         try {
             $content = $this->contentService->loadContentByRemoteId($remoteId);
-
+            if ($mode === static::MODE_INSERT_ONLY) {
+                return false;
+            }
             return ContentUpdateStructure::createForContentId($content->id, $language, $data);
         } catch (NotFoundException $e) {
             // The content doesn't exist yet, so it will be created.
         }
 
+        if ($mode === static::MODE_UPDATE_ONLY) {
+            return false;
+        }
         return new ContentCreateStructure(
             $contentType,
             $language,