+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Update to use Dataflow v2.0+
 * Add compiler pass to change the Dataflow DBAL connection factory
 * Use the DBAL connection from siteaccess
+* Add `mode` argument on `ContentStructureFactory::transform()` function
 
 # version 1.0.0